+++ CHANGELOG.rst
@@ -6,7 +6,9 @@ This document describes changes between each past release.
 7.3.2 (unreleased)
 ------------------
 
-- Nothing changed yet.
+**Internal changes**
+
+- Use json instead of ujson in storage in tests (#1255)
 
 
 7.3.1 (2017-07-03)